requirements-web.txt
flask>=3.0
pywebview>=4.4
